pyproject.toml
#
# Config file for building Home Energy Management package in python
#
# Copyright 2024 Atende Industries
# Author: Agnieszka Frąc
#


[build-system]
requires = ["setuptools>=61", "typing_extensions"]
build-backend = "setuptools.build_meta"

[project]
name = "home-energy-management"
version = "1.0.0"
description = "Home energy management: decision algorithm and end devices simulators"
requires-python = ">=3.10"
dependencies = [
  "phoenixsystems-sem >= 1.0.0"
]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
  "Programming Language :: Python"
]